requirements.txt
bcrypt==4.3.0
email-validator==2.2.0
fastapi==0.115.12
openai==1.75.0
passlib==1.7.4
psycopg[binary]==3.2.13
pydantic>=2.12.0
python-dotenv==1.1.0
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.40
uvicorn==0.34.1